Add progress notification
